Update IdentityProvider types
